chore: refactor getLocalizedPaths
